Revert "Move init_logging() to an independent module: logging_console.py"
This reverts commit 3a640699fdb723a01d0d19d9aff1d216037f51cc.
